% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLabels.R
\name{percent2}
\alias{percent2}
\title{Percent scale formatting}
\usage{
percent2(x, ...)
}
\arguments{
\item{x}{Numeric input(s)}
}
\value{
Strings formatted as signed percentages
}
\description{
Percent scale formatting
}
\examples{
percent2(c(-0.2, 0, 0.5))

}
