% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{rank.normalize}
\alias{rank.normalize}
\title{Rank-normalization}
\usage{
rank.normalize(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Normal quantile transformation of the ranks of \code{x}
}
\description{
Rank-normalization
}
