% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLabels.R
\name{scientific10}
\alias{scientific10}
\title{Scientific notation for ggplot scales}
\usage{
scientific10(x)
}
\arguments{
\item{x}{"Unformatted" numeric values}
}
\value{
Parsed `expression` objects
}
\description{
Scientific notation for ggplot scales
}
\examples{
scientific10(1.2e2)    ## expression(1 \%*\% 10^+01)
}
