% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{split.rhat}
\alias{split.rhat}
\title{Split RHat}
\usage{
\method{split}{rhat}(x)
}
\arguments{
\item{x}{Numeric matrix of MCMC samples for a parameter. Each
column indexes a separate MCMC chain and each row indexes an MCMC
sample}
}
\value{
Estimated convergence diagnostic (scalar)
}
\description{
Univariate Split \eqn{\hat{R}}{R} diagnostic for MCMC convergence
}
\references{
Vehtari et al (2021) Bayesian Analysis. 16:2. pp. 667--718.
"Rank-normalization, folding, and localization: An improved
$\hat{R}$ for assessing convergence of MCMC."
}
