% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{strrev}
\alias{strrev}
\title{Reverse characters in a string}
\usage{
strrev(s)
}
\arguments{
\item{s}{Input character string}
}
\value{
The input `s` in reverse order
}
\description{
Reverse characters in a string
}
\examples{
strrev(c("foo", "bar"))
}
