% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLabels.R
\name{thousands}
\alias{thousands}
\title{Thousands separators for ggplot scales}
\usage{
thousands(x, sep = ",")
}
\arguments{
\item{x}{Numeric input(s)}
}
\value{
Strings with thousands separators
}
\description{
Thousands separators for ggplot scales
}
\examples{
thousands(1:5 * 1000)

}
